# openapi.yaml
openapi: 3.0.0
info:
  title: Nextore API
  description: |
    API Documentation for Nextore Backend.

    ### Alur Autentikasi (Auth Flow):
    Sistem menggunakan sistem autentikasi bertingkat untuk memastikan keamanan dan penugasan terminal yang tepat:
    1. **Login (`/auth/login`)**: Pengguna memasukkan `username` dan `password`.
       - Jika pengguna hanya memiliki 1 role, role tersebut otomatis aktif.
       - Jika pengguna memiliki > 1 role, flag `requiresRoleSelection` akan bernilai `true`.
    2. **Pilih Role (`/auth/select-role`)**: Diperlukan jika `requiresRoleSelection` adalah `true`. Pengguna memilih satu role aktif untuk sesi tersebut.
    3. **Pilih POS (`/auth/select-pos`)**: Diperlukan bagi pengguna dengan role `CASHIER` atau `SUPERVISOR`. Pengguna harus memilih terminal POS yang akan digunakan sebelum dapat melakukan transaksi.

    ### Akun Pengujian (QA):
    Terdapat akun pengujian dengan akses ke **semua role** (SUPERUSER, ADMIN, SUPERVISOR, CASHIER).
    - **Username:** `qa_tester`
    - **Password:** `password123`
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://nextore.nexvibe.biz.id
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session-token
  schemas:
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
          description: "Menunjukkan apakah operasi berhasil (true) atau tidak (false)."
        message:
          type: string
          example: "Operasi berhasil dilakukan"
          description: "Pesan deskriptif mengenai hasil operasi."
...
    Pos:
      type: object
      required: [id, name, location, deviceName, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: cuid
          description: "ID unik terminal POS."
          example: "pos123456789"
        name:
          type: string
          description: "Nama register/terminal."
          example: "Cashier Register 1"
        location:
          type: string
          description: "Lokasi fisik terminal."
          example: "Lantai 1 Utama"
        deviceName:
          type: string
          description: "Nama perangkat keras yang terhubung."
          example: "POS-MCH-01"
        isActive:
          type: boolean
          description: "Status aktif terminal."
          example: true
        activeUserId:
          type: string
          format: cuid
          nullable: true
          description: "ID kasir yang sedang aktif di terminal ini."
          example: "user123456789"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /pos:
    get:
      tags:
        - POS Terminals
      summary: "Ambil daftar POS terminal"
      operationId: getPosTerminals
      description: "Mendapatkan daftar terminal POS yang tersedia untuk digunakan."
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Berhasil"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Pos"
    post:
      tags:
        - POS Terminals
      summary: "Tambah POS terminal baru"
      operationId: createPos
      description: "Mendaftarkan terminal POS baru ke dalam sistem (Admin only)."
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, location, deviceName]
              properties:
                name:
                  type: string
                  example: "Kasir Depan"
                location:
                  type: string
                  example: "Lantai 1"
                deviceName:
                  type: string
                  example: "POS-DEVICE-01"
      responses:
        "201":
          description: "Berhasil dibuat"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Pos"
  /pos/{id}:
    get:
      tags:
        - POS Terminals
      summary: "Ambil detail POS terminal"
      operationId: getPosById
      description: "Mendapatkan informasi detail mengenai satu terminal POS."
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: cuid
      responses:
        "200":
          description: "Berhasil"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Pos"
    put:
      tags:
        - POS Terminals
      summary: "Perbarui POS terminal"
      operationId: updatePos
      description: "Memperbarui informasi nama, lokasi, atau nama perangkat terminal POS (Admin only)."
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: cuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                location:
                  type: string
                deviceName:
                  type: string
      responses:
        "200":
          description: "Berhasil diperbarui"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Pos"
    delete:
      tags:
        - POS Terminals
      summary: "Hapus POS terminal"
      operationId: deletePos
      description: "Menghapus terminal POS dari sistem (Admin only)."
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: cuid
      responses:
        "200":
          description: "Berhasil dihapus"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponse"
...